Changes Summary:
Optimization enabled: We've updated foundry.toml file in order to enable optimization. We found out that optimizer_runs = 100 helps to reduce the Deployment Cost and the Deployment size of most of the functions (including Transfer function) 

How much reduce Deployment gas Cost?
This change reduce:
- Gas cost from 2675716 to 2656338 in my local environment.
- Transfer function avg from 215056 to 204984 in my local environment.

Before
```
forge test --gas-report
[⠆] Compiling...
[⠑] Compiling 20 files with 0.8.0
[⠘] Solc 0.8.0 finished in 2.44s
Compiler run successful!

Ran 16 tests for test/Gas.UnitTests.t.sol:GasTest
[PASS] testAddHistory() (gas: 233)
[PASS] testAddToWhitelist(address,uint256) (runs: 256, μ: 47734, ~: 47705)
[PASS] testBalanceOf() (gas: 12313)
[PASS] testCheckForAdmin() (gas: 19660)
[PASS] testGetPaymentHistory() (gas: 212)
[PASS] testGetPaymentStatus(address) (runs: 256, μ: 344, ~: 344)
[PASS] testGetTradingMode() (gas: 210)
[PASS] testTransfer(uint256,address) (runs: 256, μ: 234550, ~: 237970)
[PASS] testWhiteTranferAmountUpdate(address,address,uint256,string,uint256) (runs: 256, μ: 419197, ~: 419220)
[PASS] test_admins() (gas: 36302)
[PASS] test_onlyOwner(address,uint256) (runs: 256, μ: 49300, ~: 49388)
[PASS] test_tiers(address,uint256) (runs: 256, μ: 80109, ~: 79551)
[PASS] test_tiersReverts(address,uint256) (runs: 256, μ: 47410, ~: 47419)
[PASS] test_whitelistEvents(address,address,uint256,string,uint256) (runs: 256, μ: 406109, ~: 410099)
[PASS] test_whitelistEvents(address,uint256) (runs: 256, μ: 82543, ~: 81845)
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 405432, ~: 409936)
Suite result: ok. 16 passed; 0 failed; 0 skipped; finished in 334.26ms (933.38ms CPU time)
| src/Gas.sol:GasContract contract |                 |        |        |        |         |
|----------------------------------|-----------------|--------|--------|--------|---------|
| Deployment Cost                  | Deployment Size |        |        |        |         |
| 2675716                          | 12406           |        |        |        |         |
| Function Name                    | min             | avg    | median | max    | # calls |
| addToWhitelist                   | 35277           | 66252  | 83296  | 85342  | 8       |
| administrators                   | 2547            | 2547   | 2547   | 2547   | 5       |
| balanceOf                        | 660             | 2160   | 2660   | 2660   | 8       |
| balances                         | 598             | 1098   | 598    | 2598   | 4       |
| checkForAdmin                    | 12027           | 12027  | 12027  | 12027  | 1       |
| getPaymentStatus                 | 807             | 807    | 807    | 807    | 1       |
| transfer                         | 214873          | 215056 | 215095 | 215161 | 4       |
| whiteTransfer                    | 105417          | 105433 | 105441 | 105441 | 3       |
| whitelist                        | 642             | 642    | 642    | 642    | 2       |

Ran 1 test suite in 358.76ms (334.26ms CPU time): 16 tests passed, 0 failed, 0 skipped (16 total tests)
```

After
```
forge test --gas-report
[⠢] Compiling...
No files changed, compilation skipped

Ran 16 tests for test/Gas.UnitTests.t.sol:GasTest
[PASS] testAddHistory() (gas: 277)
[PASS] testAddToWhitelist(address,uint256) (runs: 256, μ: 47698, ~: 47659)
[PASS] testBalanceOf() (gas: 12267)
[PASS] testCheckForAdmin() (gas: 19614)
[PASS] testGetPaymentHistory() (gas: 189)
[PASS] testGetPaymentStatus(address) (runs: 256, μ: 321, ~: 321)
[PASS] testGetTradingMode() (gas: 276)
[PASS] testTransfer(uint256,address) (runs: 256, μ: 235154, ~: 237905)
[PASS] testWhiteTranferAmountUpdate(address,address,uint256,string,uint256) (runs: 256, μ: 419531, ~: 419573)
[PASS] test_admins() (gas: 36609)
[PASS] test_onlyOwner(address,uint256) (runs: 256, μ: 49344, ~: 49398)
[PASS] test_tiers(address,uint256) (runs: 256, μ: 80302, ~: 79582)
[PASS] test_tiersReverts(address,uint256) (runs: 256, μ: 47435, ~: 47452)
[PASS] test_whitelistEvents(address,address,uint256,string,uint256) (runs: 256, μ: 405639, ~: 410166)
[PASS] test_whitelistEvents(address,uint256) (runs: 256, μ: 82445, ~: 81898)
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 406012, ~: 410338)
Suite result: ok. 16 passed; 0 failed; 0 skipped; finished in 254.50ms (778.93ms CPU time)
| src/Gas.sol:GasContract contract |                 |        |        |        |         |
|----------------------------------|-----------------|--------|--------|--------|---------|
| Deployment Cost                  | Deployment Size |        |        |        |         |
| 2656338                          | 12316           |        |        |        |         |
| Function Name                    | min             | avg    | median | max    | # calls |
| addToWhitelist                   | 35230           | 66714  | 85216  | 85303  | 8       |
| administrators                   | 2613            | 2613   | 2613   | 2613   | 5       |
| balanceOf                        | 637             | 2137   | 2637   | 2637   | 8       |
| balances                         | 642             | 1142   | 642    | 2642   | 4       |
| checkForAdmin                    | 12004           | 12004  | 12004  | 12004  | 1       |
| getPaymentStatus                 | 873             | 873    | 873    | 873    | 1       |
| transfer                         | 194812          | 204984 | 204978 | 215169 | 4       |
| whiteTransfer                    | 105507          | 105507 | 105507 | 105507 | 3       |
| whitelist                        | 619             | 619    | 619    | 619    | 2       |

Ran 1 test suite in 257.05ms (254.50ms CPU time): 16 tests passed, 0 failed, 0 skipped (16 total tests)
```