@susumutomita 2024-03-08
Adjusting the `optimizer_runs` setting demonstrates potential for further optimization. Specifically, when `optimizer_runs` is set to 10, we observed a reduction in gas consumption from 2,412,246 to 2,398,431, indicating that lower values can lead to more efficient gas usage.
Can you verify in your side as well?

### Case optimizer_runs = 10
```
❯ forge test --gas-report
[⠢] Compiling...
[⠰] Compiling 20 files with 0.8.0
[⠔] Solc 0.8.0 finished in 3.28s
Compiler run successful!

Ran 16 tests for test/Gas.UnitTests.t.sol:GasTest
[PASS] testAddHistory() (gas: 386)
[PASS] testAddToWhitelist(address,uint256) (runs: 256, μ: 25980, ~: 25983)
[PASS] testBalanceOf() (gas: 12485)
[PASS] testCheckForAdmin() (gas: 20472)
[PASS] testGetPaymentHistory() (gas: 452)
[PASS] testGetPaymentStatus(address) (runs: 256, μ: 430, ~: 430)
[PASS] testGetTradingMode() (gas: 694)
[PASS] testTransfer(uint256,address) (runs: 256, μ: 209881, ~: 212877)
[PASS] testWhiteTranferAmountUpdate(address,address,uint256,string,uint256) (runs: 256, μ: 362589, ~: 362959)
[PASS] test_admins() (gas: 38763)
[PASS] test_onlyOwner(address,uint256) (runs: 256, μ: 28066, ~: 28178)
[PASS] test_tiers(address,uint256) (runs: 256, μ: 76086, ~: 76194)
[PASS] test_tiersReverts(address,uint256) (runs: 256, μ: 25710, ~: 25710)
[PASS] test_whitelistEvents(address,address,uint256,string,uint256) (runs: 256, μ: 350122, ~: 354684)
[PASS] test_whitelistEvents(address,uint256) (runs: 256, μ: 78613, ~: 78763)
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 350869, ~: 355059)
Suite result: ok. 16 passed; 0 failed; 0 skipped; finished in 451.53ms (1.33s CPU time)
| src/Gas.sol:GasContract contract |                 |        |        |        |         |
|----------------------------------|-----------------|--------|--------|--------|---------|
| Deployment Cost                  | Deployment Size |        |        |        |         |
| 2398431                          | 12247           |        |        |        |         |
| Function Name                    | min             | avg    | median | max    | # calls |
| addToWhitelist                   | 13816           | 58452  | 85010  | 85010  | 8       |
| administrators                   | 3053            | 3053   | 3053   | 3053   | 5       |
| balanceOf                        | 746             | 2246   | 2746   | 2746   | 8       |
| balances                         | 596             | 1096   | 596    | 2596   | 4       |
| checkForAdmin                    | 12444           | 12444  | 12444  | 12444  | 1       |
| getPaymentStatus                 | 982             | 982    | 982    | 982    | 1       |
| transfer                         | 171118          | 185582 | 190071 | 191071 | 4       |
| whiteTransfer                    | 95439           | 96772  | 97439  | 97439  | 3       |
| whitelist                        | 904             | 904    | 904    | 904    | 2       |

Ran 1 test suite in 455.58ms (451.53ms CPU time): 16 tests passed, 0 failed, 0 skipped (16 total tests)
```
 
On the other hand 
### Case optimizer_runs = 100
```
❯ forge test --gas-report
[⠢] Compiling...
[⠰] Compiling 20 files with 0.8.0
[⠔] Solc 0.8.0 finished in 3.26s
Compiler run successful!

Ran 16 tests for test/Gas.UnitTests.t.sol:GasTest
[PASS] testAddHistory() (gas: 277)
[PASS] testAddToWhitelist(address,uint256) (runs: 256, μ: 26075, ~: 26075)
[PASS] testBalanceOf() (gas: 12267)
[PASS] testCheckForAdmin() (gas: 19614)
[PASS] testGetPaymentHistory() (gas: 189)
[PASS] testGetPaymentStatus(address) (runs: 256, μ: 321, ~: 321)
[PASS] testGetTradingMode() (gas: 276)
[PASS] testTransfer(uint256,address) (runs: 256, μ: 208895, ~: 212057)
[PASS] testWhiteTranferAmountUpdate(address,address,uint256,string,uint256) (runs: 256, μ: 361175, ~: 361605)
[PASS] test_admins() (gas: 36609)
[PASS] test_onlyOwner(address,uint256) (runs: 256, μ: 27812, ~: 27961)
[PASS] test_tiers(address,uint256) (runs: 256, μ: 76169, ~: 76270)
[PASS] test_tiersReverts(address,uint256) (runs: 256, μ: 25802, ~: 25802)
[PASS] test_whitelistEvents(address,address,uint256,string,uint256) (runs: 256, μ: 350818, ~: 354120)
[PASS] test_whitelistEvents(address,uint256) (runs: 256, μ: 78257, ~: 78391)
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 350221, ~: 354206)
Suite result: ok. 16 passed; 0 failed; 0 skipped; finished in 364.43ms (1.11s CPU time)
| src/Gas.sol:GasContract contract |                 |        |        |        |         |
|----------------------------------|-----------------|--------|--------|--------|---------|
| Deployment Cost                  | Deployment Size |        |        |        |         |
| 2412246                          | 12316           |        |        |        |         |
| Function Name                    | min             | avg    | median | max    | # calls |
| addToWhitelist                   | 13862           | 58501  | 85056  | 85082  | 8       |
| administrators                   | 2613            | 2613   | 2613   | 2613   | 5       |
| balanceOf                        | 637             | 2137   | 2637   | 2637   | 8       |
| balances                         | 642             | 1142   | 642    | 2642   | 4       |
| checkForAdmin                    | 12004           | 12004  | 12004  | 12004  | 1       |
| getPaymentStatus                 | 873             | 873    | 873    | 873    | 1       |
| transfer                         | 171152          | 185616 | 190105 | 191105 | 4       |
| whiteTransfer                    | 94999           | 96332  | 96999  | 96999  | 3       |
| whitelist                        | 619             | 619    | 619    | 619    | 2       |

Ran 1 test suite in 368.16ms (364.43ms CPU time): 16 tests passed, 0 failed, 0 skipped (16 total tests)
```