@susumutomita 2024-03-08
This is an interesting phenomenon. According to the Foundry Book, the default for `optimizer_runs` is 200, yet from the perspective of gas costs, it seems there might be more optimal settings. [https://book.getfoundry.sh/reference/config/solidity-compiler#optimizer_runs](https://book.getfoundry.sh/reference/config/solidity-compiler#optimizer_runs)